Split into files and reorganize folders
