+++ jobs/Backend/Task/Services/CzechCrown/ICzechNationalBankClient.cs
@@ -1,0 +1,9 @@
+﻿using Services.CzechCrown.Models;
+
+namespace Services.CzechCrown;
+
+public interface ICzechNationalBankClient
+{
+    Task<CzkExchangeRateResponse> GetExchangeRates(DateOnly date, CancellationToken ct = default);
+    Task<CzkExchangeRateResponse> GetOtherExchangeRates(DateOnly date, CancellationToken ct = default);
+}
