+++ jobs/Backend/Task/Services/IExchangeRateProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace Services;
+
+public interface IExchangeRateProvider
+{
+    public Task<IEnumerable<ExchangeRate>> GetExchangeRates(IEnumerable<Currency> currencies, CancellationToken ct = default);
+}
