+++ jobs/Backend/Task/Services/ServiceCollectionExtension.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Polly;
+using Services.CzechCrown;
 using Services.Options;
 
 namespace Services;
